+++ src/components/CCIP/Cards/TokenCard.css
@@ -11,6 +11,7 @@
   border: 1px solid #e4e8ed;
   border-radius: var(--space-1x);
   justify-content: center;
+  cursor: pointer;
 }
 
 .token-card__container:hover {
