+++ src/components/CCIP/ChainHero/ChainHero.css
@@ -9,16 +9,33 @@
 
 .ccip-hero__heading {
   color: var(--gray-900);
-  font-size: 26px;
-  font-weight: 500;
   display: flex;
   align-items: center;
   gap: var(--space-6x);
 }
 
+.ccip-hero__heading h1 {
+  margin: 0;
+  font-size: 28px;
+  font-weight: 500;
+}
+
 .ccip-hero__heading img {
   width: 80px;
   height: 80px;
+}
+
+.ccip-hero__heading__images {
+  position: relative;
+}
+
+.ccip-hero__heading__images__small {
+  width: var(--space-10x) !important;
+  height: var(--space-10x) !important;
+  border-radius: 50%;
+  top: 60%;
+  left: 60%;
+  position: absolute;
 }
 
 .ccip-hero__content {
@@ -77,6 +94,12 @@
   font-feature-settings: "liga" off;
 
   color: #212732;
+}
+
+.ccip-hero__token-logo__symbol {
+  font-weight: 500;
+  font-size: 18px;
+  color: var(--gray-500);
 }
 
 @media (min-width: 50em) {
