+++ src/components/CCIP/Landing/ccip-landing.astro
@@ -73,7 +73,11 @@ const searchLanes = getSearchLanes({ environment })
       <div class="tokens__grid">
         {
           allTokens.map((token) => (
-            <TokenCard name={token.name} logo={token.logo} basePath={`/ccip/supported-networks/${environment}`} />
+            <TokenCard
+              name={token.name}
+              logo={token.logo}
+              link={`/ccip/supported-networks/${environment}/token/${token.name}`}
+            />
           ))
         }
       </div>
