+++ src/components/CCIP/Tables/Table.css
@@ -89,6 +89,7 @@
 .ccip-table__filters {
   display: flex;
   justify-content: space-between;
+  align-items: center;
   gap: 16px;
   margin-bottom: var(--space-6x);
   border-bottom: 1px solid var(--gray-200);
