+++ src/components/CCIP/Token/Token.astro
@@ -59,8 +59,9 @@ const searchLanes = getSearchLanes({ environment })
     client:load
     lanes={searchLanes}
     token={{
-      name: token || "",
+      name: data[Object.keys(data)[0]]?.name || "",
       logo: logo || "",
+      symbol: token || "",
     }}
     environment={environment}
   />
@@ -92,8 +93,9 @@ const searchLanes = getSearchLanes({ environment })
         })}
         lanes={lanes}
         token={{
-          name: token || "",
+          name: data[Object.keys(data)[0]]?.name || "",
           logo: logo || "",
+          symbol: token,
         }}
       />
     </div>
